+++ src/app/pockets/complete/page.tsx
@@ -6,7 +6,7 @@ import { kakaotalkShare } from "@/utils/share/kakaotalk-share";
 import ShareModal from "@/components/modal/share-modal";
 import { generateUrl } from "@/utils/url/urlGenerator";
 import usePocketStore from "@/store/pocket";
-import { pocketsImageData } from "@/utils/images/cardNames";
+import { pocketsImageData } from "@/data/card-names";
 import Header from "@/components/header/header";
 import BackButton from "@/components/header/back-button";
 import { buttonBackClick } from "@/utils/edit/buttonBackClick";
