--- src/utils/images/image.ts
@@ -1,28 +1,0 @@
-// 이미지 불러오는 유틸 함수
-
-export const pocketImages = [
-  "pocketsImage/복주머니-01.png",
-  "pocketsImage/복주머니-02.png",
-  "pocketsImage/복주머니-03.png",
-  "pocketsImage/복주머니-04.png",
-  "pocketsImage/복주머니-05.png",
-  "pocketsImage/복주머니-06.png",
-  "pocketsImage/복주머니-07.png",
-  "pocketsImage/복주머니-08.png",
-];
-
-export const cardImages = [
-  "/card/FORTUNETOSS_CARD_1.webp",
-  "/card/FORTUNETOSS_CARD_2.webp",
-  "/card/FORTUNETOSS_CARD_3.webp",
-  "/card/FORTUNETOSS_CARD_4.webp",
-  "/card/FORTUNETOSS_CARD_5.webp",
-];
-
-export const cardBackImages = [
-  "/cardBack/FORTUNETOSS_CARD_BACK_1.webp",
-  "/cardBack/FORTUNETOSS_CARD_BACK_2.webp",
-  "/cardBack/FORTUNETOSS_CARD_BACK_3.webp",
-  "/cardBack/FORTUNETOSS_CARD_BACK_4.webp",
-  "/cardBack/FORTUNETOSS_CARD_BACK_5.webp",
-];
